requirements.txt
requests
dateparser
python-dateutil
rich
